feat: EDM4eic file merger (background treatment) (#3)
### Briefly, what does this PR introduce?
We want a small utility to merge background hits into main production
simulations. This is possibly another way to achieve this without having
to depend on the event reconstruction framework. It would be an
'afterburner' after the simulation.

This utility takes at a minimum two arguments: a signal file and a
background file. It adds one events from the background file for each
event in the signal file.

### What kind of change does this PR introduce?
- [ ] Bug fix (issue #__)
- [X] New feature (issue #__)
- [ ] Documentation update
- [ ] Other: __

### Please check if this PR fulfills the following:
- [ ] Tests for the changes have been added
- [ ] Documentation has been added / updated
- [ ] Changes have been communicated to collaborators

### Does this PR introduce breaking changes? What changes might users
need to make to their code?
No breaking changes, just an additional utility.

### Does this PR change default behavior?
No changes to default behavior.