feat: add version header (#40)
### Briefly, what does this PR introduce?
This adds a version header to the installation of EDM4eic. This allows
us to make downstream code behave differently based on the version.

Note: with `.gitattributes` (in particular `export-subst`) we could
automate the filling of this version information with the most recent
tag and such. Maybe something to look into.

---------

Co-authored-by: Dmitry Kalinkin <dmitry.kalinkin@gmail.com>