39 add surface id member to edm4eictrackpoint (#41)
### Briefly, what does this PR introduce?
This adds two data members to TrackPoint to identify the detector
and surface on which the TrackPoint lies (implemented for track
projections but could be used more generally).

This change replaces the original change which added a single 32-bit
integer. This was done following a discussion in the S&C meeting on
August 23, 2023. Summary of proposed options can be [found
here](https://indico.bnl.gov/event/20352/contributions/80021/attachments/49313/84184/followup_edm4eic_trackpoint_PR%20copy.pdf).

### What kind of change does this PR introduce?
- [x] New feature (issue #39)