remove simID and recID from the associations
simID was originally introduced in eicd as there was no way to refer to a
EDM4hep collection at the time
https://github.com/eic/EDM4eic/blame/7299aba54097a3f2a17c0242731c9dabe68fb450/eic_data.yaml#L546-L547
Later, recID was introduced for convenience, and possibly had to be used to
hack associations into an early versions of EICrecon.  We don't need neither.
This information is already available via sim and rec relations.