feat: Measurement2D to represent 2D surface measurements (#54)
### Briefly, what does this PR introduce?
This adds a new datatype, `Measurement2D`, intended primarily but not
exclusively for use in Acts track finding. Therefore it tends to mirror
what we are already doing. The datatype stores two-dimensional (+time)
constraints on a surface with identifier.


There is additional motivation and discussion at
https://docs.google.com/presentation/d/1ee_iRAyeUao7PwkmFnapX7cASAp4rm8QnQ_moidYDnk/edit,
but changes may be applied here without updating the slides.

### What kind of change does this PR introduce?
- [ ] Bug fix (issue #__)
- [x] New feature (issue #__)
- [ ] Documentation update
- [ ] Other: __

### Please check if this PR fulfills the following:
- [ ] Tests for the changes have been added
- [ ] Documentation has been added / updated
- [x ] Changes have been communicated to collaborators

### Does this PR introduce breaking changes? What changes might users
need to make to their code?
No. Addition of new datatype.

### Does this PR change default behavior?
No.

---------

Co-authored-by: ShujieL <ShujieL@users.noreply.github.com>