fix: only compile edm4eic_merge for podio < 0.17.4 (#64)
### Briefly, what does this PR introduce?
With podio 0.17.4, we can't compile emd4eic_merge anymore since it is
reliant on podio/EventStore.h. We have been ignoring the deprecation
warnings about this for half a year now...

This PR makes the compilation of edm4eic_merge conditional on being
before that version. It renames the source and executable to
event_merge.cpp and edm4eic_event_merge to make clear this isn't working
for frames. Code is not removed since an adventurous soul may port this
to frames so we can do background mixing.

### What kind of change does this PR introduce?
- [x] Bug fix (issue: edm4eic doens't compile with podio 0.17.4)
- [ ] New feature (issue #__)
- [ ] Documentation update
- [ ] Other: __

### Please check if this PR fulfills the following:
- [ ] Tests for the changes have been added
- [ ] Documentation has been added / updated
- [ ] Changes have been communicated to collaborators

### Does this PR introduce breaking changes? What changes might users
need to make to their code?
No.

### Does this PR change default behavior?
Yes, edm4eic_merge is renamed, and will not be installed when newer
versions of podio are used.