edm4eic.yaml: fix MCRecoTrackerHitAssociation to relate to a single sim hit at a time (#77)
I believe, having sim hits as a OneToMany relation was a mistake. The
idea behind having a weight is to specify level of contribution of a
specific sim hit. If multiple sim hits are associated, user should
expect multiple association records in the collection. We are only now
introducing the actual use of this, so now is the good time to do a fix
https://github.com/eic/EICrecon/pull/1392#discussion_r1572964045

cc @wdconinc @c-dilks