Add track-protocluster link (#108)
### Briefly, what does this PR introduce?

This PR introduces the `edm4eic::TrackProtoClusterLink`, a parallel to
the existing `edm4eic::TrackClusterMatch`. This functionality was
identified as missing during discussion of
[EICrecon#1669](https://github.com/eic/EICrecon/pull/1699): in order to
ensure that merged/split clusters are reconstructed consistently with
other clusters it's preferable for the `TrackClusterMergeSplitter` to
output `edm4eic::ProtoCluster`s, but the connections between the
merged/split protoclusters and their antecedent tracks will need to be
propagated to downstream particle flow algorithms. This new type will
facilitate this.

### What kind of change does this PR introduce?
- [ ] Bug fix (issue #__)
- [X] New feature (issue #107 )
- [ ] Documentation update
- [ ] Other: __

### Please check if this PR fulfills the following:
- [ ] Tests for the changes have been added
- [ ] Documentation has been added / updated
- [X] Changes have been communicated to collaborators

### Does this PR introduce breaking changes? What changes might users
need to make to their code?

No.

### Does this PR change default behavior?

No.

---------

Co-authored-by: Dmitry Kalinkin <dmitry.kalinkin@gmail.com>